+++ apiserver/paasng/paasng/platform/modules/views.py
@@ -33,21 +33,22 @@ from rest_framework.response import Response
 
 from paas_wl.infras.cluster.shim import get_application_cluster
 from paas_wl.workloads.images.models import AppUserCredential
-from paasng.platform.bk_lesscode.client import make_bk_lesscode_client
-from paasng.platform.bk_lesscode.exceptions import LessCodeApiError, LessCodeGatewayServiceError
-from paasng.infras.iam.permissions.resources.application import AppAction
+from paasng.accessories.publish.market.models import MarketConfig
+from paasng.accessories.publish.market.protections import ModulePublishPreparer
+from paasng.core.region.models import get_region
 from paasng.infras.accounts.constants import AccountFeatureFlag as AFF
 from paasng.infras.accounts.models import AccountFeatureFlag
 from paasng.infras.accounts.permissions.application import application_perm_class, check_application_perm
-from paasng.platform.templates.constants import TemplateType
-from paasng.platform.templates.models import Template
-from paasng.platform.engine.configurations.image import generate_image_repository
-from paasng.platform.engine.constants import RuntimeType
+from paasng.infras.iam.permissions.resources.application import AppAction
 from paasng.platform.applications.mixins import ApplicationCodeInPathMixin
 from paasng.platform.applications.models import Application
 from paasng.platform.applications.signals import application_default_module_switch, pre_delete_module
 from paasng.platform.applications.specs import AppSpecs
 from paasng.platform.applications.utils import delete_module
+from paasng.platform.bk_lesscode.client import make_bk_lesscode_client
+from paasng.platform.bk_lesscode.exceptions import LessCodeApiError, LessCodeGatewayServiceError
+from paasng.platform.engine.configurations.image import generate_image_repository
+from paasng.platform.engine.constants import RuntimeType
 from paasng.platform.modules.constants import DeployHookType, SourceOrigin
 from paasng.platform.modules.exceptions import BPNotFound
 from paasng.platform.modules.helpers import (
@@ -74,9 +75,8 @@ from paasng.platform.modules.serializers import (
     ModuleSLZ,
 )
 from paasng.platform.modules.specs import ModuleSpecs
-from paasng.core.region.models import get_region
-from paasng.accessories.publish.market.models import MarketConfig
-from paasng.accessories.publish.market.protections import ModulePublishPreparer
+from paasng.platform.templates.constants import TemplateType
+from paasng.platform.templates.models import Template
 from paasng.utils.api_docs import openapi_empty_response
 from paasng.utils.error_codes import error_codes
 from paasng.utils.views import permission_classes as perm_classes
@@ -290,6 +290,7 @@ class ModuleViewSet(viewsets.ViewSet, ApplicationCodeInPathMixin):
             source_dir=source_config.get('source_dir', ''),
             cluster_name=cluster.name,
             manifest=data.get('manifest'),
+            build_config=serializer.validated_data['build_config'],
         )
 
         return Response(
